evo_trans/AdaIN/requirements.txt
numpy==1.19.5
Pillow==8.4.0
protobuf==3.15.0
six==1.12.0
tensorboardX==1.8
torch==1.2.0
torchvision==0.4.0
tqdm==4.35.0
opencv-python==4.4.0.46
imageio==2.9.0
